Import from HNS - signature change
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/simplewebkit/trunk@38155 72102866-910b-0410-8b05-ffd578937521

+++ Sources/WebBackForwardList.m
@@ -75,7 +75,7 @@
 
 - (void) goToItem:(WebHistoryItem *) item;
 {
-	unsigned idx;
+	NSUInteger idx;
 	if(item == _currentItem)
 		return;	// already there!
 	idx=[_backList indexOfObjectIdenticalTo:item];	// is it here?
